Initial migration of overlay conditions to Resource Loader API
