--- fabric-overlay-conditions-api-v1/build.gradle
@@ -1,6 +1,0 @@
-version = getSubprojectVersion(project)
-
-testDependencies(project, [
-	':fabric-lifecycle-events-v1',
-	':fabric-resource-loader-v0'
-])
