--- fabric-overlay-conditions-api-v1/src/main/java/net/fabricmc/fabric/api/overlay/conditions/v1/package-info.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Provides a way of conditionally applying resource pack overlays.
- * Conditions are identified by an identifier and registered at {@link
- * net.fabricmc.fabric.api.overlay.conditions.v1.OverlayConditions}.
- *
- * <h2>JSON format</h2>
- *
- * <p>Add an array with the {@code fabric:overlays} key to the pack.mcmeta file:
- * <pre>{@code
- * {
- *   "pack": {
- *     "pack_format": 26,
- *     "description": ""
- *   },
- *   "fabric:overlays": {
- *       "entries": [
- *  *     {
- *  *       "directory": "another_folder",
- *  *       "type": "<insert condition type ID here>",
- *  *       // values of the condition
- *  *     }
- *  *   ]
- *   }
- * }
- * }</pre>
- *
- * <p>See {@link net.fabricmc.fabric.api.overlay.conditions.v1.DefaultOverlayConditions} for
- * the list of built-in conditions. It is also possible to register custom conditions.
- */
-package net.fabricmc.fabric.api.overlay.conditions.v1;
