--- fabric-overlay-conditions-api-v1/src/main/resources/fabric-overlay-conditions-api-v1.mixins.json
@@ -1,11 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.overlay.conditions",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "ResourcePackProfileMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
