--- fabric-overlay-conditions-api-v1/src/main/resources/fabric.mod.json
@@ -1,28 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-overlay-conditions-api-v1",
-  "name": "Fabric Overlay Conditions API (v1)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-overlay-conditions-api-v1/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.15.1"
-  },
-  "description": "Allows conditionally loading resource pack overlays.",
-  "mixins": [
-    "fabric-overlay-conditions-api-v1.mixins.json"
-  ],
-  "custom": {
-    "fabric-api:module-lifecycle": "stable"
-  }
-}
