--- fabric-overlay-conditions-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,18 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-overlay-conditions-api-v1-testmod",
-  "name": "Fabric Overlay Conditions API (v1) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-overlay-conditions-api-v1": "*",
-    "fabric-resource-loader-v0": "*",
-    "fabric-lifecycle-events-v1": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.test.overlay.conditions.OverlayConditionsTest"
-    ]
-  }
-}
