+++ fabric-resource-loader-v0/src/main/java/net/fabricmc/fabric/api/resource/OverlayConditions.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.overlay.conditions.v1;
+package net.fabricmc.fabric.api.resource;
 
 import java.util.Map;
 import java.util.Objects;
