+++ fabric-resource-loader-v0/src/testmod/resources/resourcepacks/overlay_test/do_overlay/data/fabric-resource-loader-v0-testmod/recipes/do_overlay.json
