+++ fabric-resource-loader-v0/src/testmod/resources/resourcepacks/overlay_test/dont_overlay/data/fabric-resource-loader-v0-testmod/recipes/dont_overlay.json
