+++ fabric-resource-loader-v0/src/testmod/resources/resourcepacks/overlay_test/pack.mcmeta
@@ -8,14 +8,14 @@
       {
         "directory": "do_overlay",
         "type": "fabric:mod_loaded",
-        "mod_id": "fabric-overlay-conditions-api-v1"
+        "mod_id": "fabric-resource-loader-v0-testmod"
       },
       {
         "directory": "dont_overlay",
         "type": "fabric:not",
         "condition": {
           "type": "fabric:mod_loaded",
-          "mod_id": "fabric-overlay-conditions-api-v1"
+          "mod_id": "fabric-resource-loader-v0-testmod"
         }
       }
     ]
