+++ gradle.properties
@@ -44,7 +44,6 @@ fabric-model-loading-api-v1-version=1.0.8
 fabric-models-v0-version=0.4.7
 fabric-networking-api-v1-version=3.1.7
 fabric-object-builder-api-v1-version=13.0.11
-fabric-overlay-conditions-api-v1-version=1.0.0
 fabric-particles-v1-version=1.1.7
 fabric-recipe-api-v1-version=2.0.20
 fabric-registry-sync-v0-version=4.0.18
