+++ settings.gradle
@@ -43,7 +43,6 @@ include 'fabric-mining-level-api-v1'
 include 'fabric-model-loading-api-v1'
 include 'fabric-networking-api-v1'
 include 'fabric-object-builder-api-v1'
-include 'fabric-overlay-conditions-api-v1'
 include 'fabric-particles-v1'
 include 'fabric-recipe-api-v1'
 include 'fabric-registry-sync-v0'
